Fix missing directory_id in ResolveNames response
Parsing of the ResolveNames response checked if a Contact element
existed. If so, it also expected it to have a DirectoryId element of
the contact. But the Contact element can be filled with information
about a resolved mailbox. Which in turn does not have a DirectoryId
element.
In this case, the Contact element will be ignored now.